requirements.txt
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
accelerate=1.0.1=pypi_0
addict=2.4.0=pypi_0
aiohappyeyeballs=2.4.4=pypi_0
aiohttp=3.10.11=pypi_0
aiosignal=1.3.1=pypi_0
antlr4-python3-runtime=4.8=pypi_0
appdirs=1.4.4=pypi_0
asttokens=3.0.0=pypi_0
async-timeout=5.0.1=pypi_0
attrs=25.3.0=pypi_0
audioread=3.0.1=pypi_0
backcall=0.2.0=pypi_0
bitarray=3.4.0=pypi_0
blinker=1.8.2=pypi_0
braceexpand=0.1.7=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.2.25=h06a4308_0
cachetools=5.5.2=pypi_0
certifi=2025.6.15=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.4.2=pypi_0
click=8.1.3=pypi_0
comm=0.2.2=pypi_0
configargparse=1.7=pypi_0
contourpy=1.1.1=pypi_0
cycler=0.12.1=pypi_0
cython=3.1.0=pypi_0
dash=3.0.2=pypi_0
datasets=3.1.0=pypi_0
decorator=5.2.1=pypi_0
dill=0.3.8=pypi_0
docker-pycreds=0.4.0=pypi_0
easydict=1.10=pypi_0
einops=0.8.1=pypi_0
exceptiongroup=1.3.0=pypi_0
executing=2.2.0=pypi_0
expat=2.7.1=h6a678d5_0
faiss-gpu=1.7.2=pypi_0
fastjsonschema=2.21.1=pypi_0
filelock=3.16.1=pypi_0
flask=3.0.3=pypi_0
fonttools=4.57.0=pypi_0
frozenlist=1.5.0=pypi_0
fsspec=2024.9.0=pypi_0
ftfy=6.1.1=pypi_0
gitdb=4.0.12=pypi_0
gitpython=3.1.44=pypi_0
h5pickle=0.4.2=pypi_0
h5py=3.11.0=pypi_0
hf-xet=1.1.5=pypi_0
huggingface-hub=0.33.2=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.7.0=pypi_0
importlib-resources=6.4.5=pypi_0
iniconfig=2.1.0=pypi_0
inquirerpy=0.3.4=pypi_0
ipython=8.12.3=pypi_0
ipywidgets=8.1.5=pypi_0
itsdangerous=2.2.0=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.6=pypi_0
joblib=1.4.2=pypi_0
jsonlines=4.0.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter-core=5.8.1=pypi_0
jupyterlab-widgets=3.0.13=pypi_0
kiwisolver=1.4.7=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
librosa=0.10.1=pypi_0
libstdcxx-ng=11.2.0=h1234567_1
libxcb=1.17.0=h9b100fa_0
llvmlite=0.41.1=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.7.5=pypi_0
matplotlib-inline=0.1.7=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.1.0=pypi_0
multidict=6.1.0=pypi_0
multiprocess=0.70.16=pypi_0
narwhals=1.33.0=pypi_0
nbformat=5.7.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pypi_0
networkx=3.1=pypi_0
ninja=1.11.1=pypi_0
numba=0.58.1=pypi_0
numpy=1.24.4=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.20.5=pypi_0
nvidia-nvjitlink-cu12=12.9.86=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
open-clip-torch=2.32.0=pypi_0
open3d=0.17.0=pypi_0
opencv-python=4.7.0.72=pypi_0
openssl=3.0.16=h5eee18b_0
packaging=25.0=pypi_0
pandas=2.0.1=pypi_0
parso=0.8.4=pypi_0
pathtools=0.1.2=pypi_0
pexpect=4.9.0=pypi_0
pfzy=0.3.4=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.5.0=pypi_0
pip=25.1=pyhc872135_2
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=4.3.8=pypi_0
plotly=6.0.1=pypi_0
pluggy=1.5.0=pypi_0
pooch=1.8.2=pypi_0
portalocker=3.0.0=pypi_0
progressbar=2.5=pypi_0
prompt-toolkit=3.0.50=pypi_0
propcache=0.2.0=pypi_0
protobuf=4.25.7=pypi_0
psutil=7.0.0=pypi_0
pthread-stubs=0.3=h0ce48e5_1
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.3=pypi_0
pyarrow=17.0.0=pypi_0
pycparser=2.22=pypi_0
pygments=2.19.1=pypi_0
pynndescent=0.5.13=pypi_0
pyparsing=3.1.4=pypi_0
pyquaternion=0.9.9=pypi_0
pytest=8.3.5=pypi_0
python=3.9.23=he99959a_0
python-dateutil=2.9.0.post0=pypi_0
pytz=2025.2=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.4=pypi_0
retrying=1.3.4=pypi_0
rpds-py=0.20.1=pypi_0
sacrebleu=1.5.0=pypi_0
safetensors=0.5.3=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.10.1=pypi_0
seaborn=0.13.2=pypi_0
sentence-transformers=3.2.1=pypi_0
sentencepiece=0.2.0=pypi_0
sentry-sdk=2.25.1=pypi_0
setproctitle=1.3.5=pypi_0
setuptools=78.1.1=py39h06a4308_0
six=1.17.0=pypi_0
smmap=5.0.2=pypi_0
soundfile=0.12.1=pypi_0
soxr=0.3.7=pypi_0
sqlite=3.50.2=hb25bd0a_1
stack-data=0.6.3=pypi_0
sympy=1.13.3=pypi_0
tensorboardx=2.6.2.2=pypi_0
termcolor=2.3.0=pypi_0
threadpoolctl=3.5.0=pypi_0
timm=1.0.17=pypi_0
tk=8.6.14=h993c535_1
tokenizers=0.21.2=pypi_0
tomli=2.2.1=pypi_0
torch=2.1.0+cu121=pypi_0
torchaudio=2.1.0+cu121=pypi_0
torchvision=0.16.0+cu121=pypi_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pypi_0
transformers=4.53.2=pypi_0
triton=2.1.0=pypi_0
typing-extensions=4.13.2=pypi_0
tzdata=2025.2=pypi_0
umap=0.1.1=pypi_0
umap-learn=0.5.7=pypi_0
urllib3=2.2.3=pypi_0
wandb=0.15.12=pypi_0
wcwidth=0.2.13=pypi_0
webdataset=0.2.100=pypi_0
werkzeug=3.0.6=pypi_0
wget=3.2=pypi_0
wheel=0.45.1=py39h06a4308_0
widgetsnbextension=4.0.13=pypi_0
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xxhash=3.5.0=pypi_0
xz=5.6.4=h5eee18b_1
yarl=1.15.2=pypi_0
zipp=3.20.2=pypi_0
zlib=1.2.13=h5eee18b_1
